+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Sequencer/TargetSchedulerSyncContainer.cs
@@ -22,7 +22,7 @@ namespace Assistant.NINAPlugin.Sequencer {
     [ExportMetadata("Name", "Target Scheduler Sync Container")]
     [ExportMetadata("Description", "Target Scheduler synchronized imaging for multiple NINA instances")]
     [ExportMetadata("Icon", "Scheduler.SchedulerSVG")]
-    [ExportMetadata("Category", "Lbl_SequenceCategory_Container")]
+    [ExportMetadata("Category", "Target Scheduler")]
     [Export(typeof(ISequenceItem))]
     [Export(typeof(ISequenceContainer))]
     [JsonObject(MemberSerialization.OptIn)]