+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Target Scheduler
 
-## 4.0.1.0 - 2023-11-XX
+## 4.0.2.0 - 2023-11-XX
 * Added image grading on FWHM and Eccentricity (requires Hocus Focus plugin)
 * Added option to move rejected images to a 'rejected' directory
 * Added ability to purge acquired image records by date or date/target
 * Added smarter determination of plan stop times
+* Added ability to delete all target exposure plans
 * The rule weight list is now sorted when displayed
 * Added target rotation and ROI to the set of data saved for acquired images.  A future release will use these values when selecting 'like' images for grading.
 * Fixed issue where target rotation wasn't being sent to Framing Wizard